// src/auto-generated/api-configs/enums.ts
export enum Actions {
  LOGIN = "login",
  GET_METADATA = "get-metadata",
  GET_VERSION = "get-version",
  UPDATE_UNITS = "update-units",
  GET_MESSAGES = "get-messages",
  GET_ALL_MESSAGE_TEMPLATES = "get-all-message-templates",
  ADD_MESSAGE_TEMPLATE = "add-message-template",
  DISABLE_MESSAGE_TEMPLATE = "disable-message-template",
  ENABLE_MESSAGE_TEMPLATE = "enable-message-template",
  GET_USERS = "get-users",
  CHANGE_PASSWORD = "change-password",
  DISABLE_USERS = "disable-users",
  ADD_USER = "add-user",
  UPDATE_USER = "update-user",
  GET_DEPARTMENTS = "get-departments",
  ADD_DEPARTMENT = "add-department",
  UPDATE_DEPARTMENT = "update-department",
  DELETE_DEPARTMENT = "delete-department",
  GET_CUSTOMERS = "get-customers",
  UPDATE_CUSTOMER = "update-customer",
  GET_CUSTOMER_PRODUCTS = "get-customer-products",
  UPDATE_CUSTOMER_PRODUCT = "update-customer-product",
  GET_PRODUCTS = "get-products",
  GET_ALL_PRODUCTS = "get-all-products",
  ADD_PRODUCT = "add-product",
  // UPDATE_PRODUCT = "update-product",
  // DELETE_PRODUCT = "delete-product",
  // GET_MENUS = "get-menus",
  GET_BOM = "get-bom",
  PUSH_BOM = "push-bom",
  ADD_MENU = "add-menu",
  GET_DAILY_MENU = "get-daily-menu",
  PUSH_DAILY_MENU = "push-daily-menu",
  GET_TODAY_MENU = "get-today-menu",
  // UPDATE_MENU = "update-menu",
  // DELETE_MENU = "delete-menu",
  GET_ALL_MATERIALS = "get-all-materials",
  GET_MATERIALS = "get-materials",
  PUSH_MATERIAL = "push-material",
  UPDATE_MATERIAL_SUPPLIER = "update-material-supplier",
  GET_INVENTORY = "get-inventory",
  UPDATE_INVENTORY = "update-inventory",
  GET_LOW_INVENTORIES = "get-low-inventories",
  GET_PERIODIC_INVENTORIES = "get-periodic-inventories",
  GET_INVENTORIES_FOR_DAILY_MENU = "get-inventories-for-daily-menu",
  GET_MATERIAL_INVENTORIES = "get-material-inventories",
  GET_MONTHLY_INVENTORIES = "get-monthly-inventories",
  GET_SUPPLIERS = "get-suppliers",
  ADD_SUPPLIER = "add-supplier",
  UPDATE_SUPPLIER = "update-supplier",
  UPDATE_SUPPLIER_MATERIAL = "update-supplier-material",
  GET_PURCHASE_REQUESTS = "get-purchase-requests",
  ADD_PURCHASE_REQUEST = "add-purchase-request",
  UPDATE_PURCHASE_REQUEST = "update-purchase-request",
  GET_PURCHASE_INTERNALS = "get-purchase-internals",
  ADD_PURCHASE_INTERNAL = "add-purchase-internal",
  UPDATE_PURCHASE_INTERNAL = "update-purchase-internal",
  UPDATE_PURCHASE_INTERNAL_STATUS = "update-purchase-internal-status",
  GET_PURCHASE_COORDINATIONS = "get-purchase-coordinations",
  ADD_PURCHASE_COORDINATION = "add-purchase-coordination",
  UPDATE_PURCHASE_COORDINATION = "update-purchase-coordination",
  GET_PURCHASE_ORDERS = "get-purchase-orders",
  ADD_PURCHASE_ORDER = "add-purchase-order",
  UPDATE_PURCHASE_ORDER = "update-purchase-order",
  UPDATE_PURCHASE_ORDER_STATUS = "update-purchase-order-status",
  GET_WAREHOUSE_RECEIPTS = "get-warehouse-receipts",
  GET_WAREHOUSE_EXPORTS = "get-warehouse-exports",
  GET_WAREHOUSE_IMPORTS = "get-warehouse-imports",
}